Add Windows support for the CLI application.  Closes #11.
composer handles windows support in general[1], but depending on how
this repository is being accessed (e.g., if it is cloned from github),
then composer may not be run to create the which.bat script.  This
which.bat script serves as a simple wrapper around the which command for
all use cases.

[1]: https://getcomposer.org/doc/articles/vendor-binaries.md#what-about-windows-and-bat-files-